fix(core): normalize input for poll option
This change normalizes input for the poll option to either a boolean or number to meet the specs required by Webpack config.
Move the function into the docusaurus.mjs file as requested.

fixes #8306
